Moved GetVerificationEmailContentFn and GetPasswordResetEmailContentFn to SDK.
