+++ waspc/data/Generator/templates/sdk/auth/user.ts
@@ -2,7 +2,7 @@
 // We have them duplicated in this file and in data/Generator/templates/server/src/auth/user.ts
 // If you are changing the logic here, make sure to change it there as well.
 
-import type { User, ProviderName, DeserializedAuthIdentity } from './types' 
+import type { User, ProviderName, DeserializedAuthIdentity } from './types'
 
 export function getEmail(user: User): string | null {
   return findUserIdentity(user, "email")?.providerUserId ?? null;
