+++ waspc/data/Generator/templates/sdk/package.json
@@ -108,6 +108,8 @@
       "./server/actions": "./dist/server/actions/index.js",
       {=!  Used by our code, uncodumented (but accessible) for users. =}
       "./server/queries": "./dist/server/queries/index.js",
+      {=!  Used by users and also by our code, documented. =}
+      "./server/auth/email": "./dist/server/auth/email/index.js",
       {=!  Used by users, documented. =}
       "./dbSeed/types": "./dist/dbSeed/types.js"
    },
