+++ waspc/data/Generator/templates/sdk/server/auth/email/index.ts
@@ -1,0 +1,9 @@
+export type GetVerificationEmailContentFn = (params: { verificationLink: string }) => EmailContent;
+
+export type GetPasswordResetEmailContentFn = (params: { passwordResetLink: string }) => EmailContent;
+
+type EmailContent = {
+    subject: string;
+    html: string;
+    text: string;
+}
