+++ waspc/data/Generator/templates/server/src/auth/providers/config/email.ts
@@ -9,7 +9,7 @@ import { getSignupRoute } from "../email/signup.js";
 import { getRequestPasswordResetRoute } from "../email/requestPasswordReset.js";
 import { resetPassword } from "../email/resetPassword.js";
 import { verifyEmail } from "../email/verifyEmail.js";
-import { GetVerificationEmailContentFn, GetPasswordResetEmailContentFn } from "../email/types.js";
+import { GetVerificationEmailContentFn, GetPasswordResetEmailContentFn } from "wasp/server/auth/email";
 import { handleRejection } from "wasp/server/utils";
 
 {=# userSignupFields.isDefined =}
