+++ waspc/data/Generator/templates/server/src/auth/providers/email/requestPasswordReset.ts
@@ -12,7 +12,7 @@ import {
 } from "./utils.js";
 import { ensureValidEmail } from 'wasp/auth/validation';
 import type { EmailFromField } from '../../../email/core/types.js';
-import { GetPasswordResetEmailContentFn } from './types.js';
+import { GetPasswordResetEmailContentFn } from 'wasp/server/auth/email';
 import HttpError from 'wasp/core/HttpError';
 
 export function getRequestPasswordResetRoute({
