+++ waspc/data/Generator/templates/server/src/auth/providers/email/types.ts
@@ -1,13 +1,3 @@
-export type GetVerificationEmailContentFn = (params: { verificationLink: string }) => EmailContent;
-
-export type GetPasswordResetEmailContentFn = (params: { passwordResetLink: string }) => EmailContent;
-
-type EmailContent = {
-    subject: string;
-    html: string;
-    text: string;
-}
-
 export const tokenVerificationErrors = {
     TokenExpiredError: 'TokenExpiredError',
 };
