--- waspc/data/Generator/templates/server/src/types/index.ts
@@ -1,5 +1,0 @@
-{{={= =}=}}
-
-{=# isEmailAuthEnabled =}
-export type { GetVerificationEmailContentFn, GetPasswordResetEmailContentFn } from '../auth/providers/email/types';
-{=/ isEmailAuthEnabled =}
