+++ waspc/src/Wasp/Generator/SdkGenerator/Auth/EmailAuthG.hs
@@ -26,6 +26,7 @@ genEmailAuth auth
         [ genIndex
         ]
         <++> genActions
+        <++> genServer
   | otherwise = return []
 
 genIndex :: Generator FileDraft
@@ -38,6 +39,12 @@ genActions =
       genSignupAction,
       genPasswordResetActions,
       genVerifyEmailAction
+    ]
+
+genServer :: Generator [FileDraft]
+genServer =
+  return
+    [ C.mkTmplFd [relfile|server/auth/email/index.ts|]
     ]
 
 genLoginAction :: Generator FileDraft
